[Decomp][ObjectMgr] Cohesion split: extract item loaders to ObjectMgrItems.cpp
